Delete unused files and change name of files to original names.
