+++ medTracker/View/MainView.swift
@@ -23,8 +23,7 @@ struct MainView: View {
             }
             // View for the users to see their information and edit it.
             NavigationStack {
-//                ProfileView(user: user, symptoms: symptoms, createAction: user.makeCreateAction(), createAction2: {_ in })
-                ProfileView2(user: user, formViewModel: user.updateUserViewModel(for: user.user))
+                ProfileView(user: user, formViewModel: user.updateUserViewModel(for: user.user))
             }
             .tabItem {
                 Label("Perfil", systemImage: "person.crop.circle")
